Moved AutoDL to a separate script
